requirements_dev.txt
# base
six
redis
wheel
twine
PyYAML
toml
flask>=1.0
python-box
python-dotenv
setuptools>=38.6.0
configobj

# testing
coveralls
pytest
pytest-cov
pytest-xdist
logzero

# style check
flake8
pep8-naming
flake8-debugger
flake8-print
flake8-todo
radon

# editable dynaconf
--editable .
